bluehack/mac_requirements.txt
bleak==0.22.3
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
kiwisolver==1.4.8
matplotlib==3.10.0
mplcursors==0.6
nest-asyncio==1.6.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyobjc-core==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-CoreBluetooth==10.3.2
pyobjc-framework-libdispatch==10.3.2
pyparsing==3.2.1
pyperclip==1.9.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
six==1.17.0
tzdata==2025.1
